+++ lib/src/typedefs.dart
@@ -4,54 +4,6 @@
 import 'dart:async';
 import 'package:text_analysis/text_analysis.dart';
 
-/// A callback function that returns true or false based on the content of
-/// the term.
-
-/// `TextSplitter` - A splitter function that returns a list of paragraphs from [source].
-///
-/// In English, the [source] text is split at:
-/// - `U+000A`, NewLine;
-/// - `U+000B`, VerticalTabulation;
-/// - `U+000C`, FormFeed; and
-/// - `U+000D`, CarriageReturn.
-
-/// `TermFilter` - A filter function that returns a collection of terms from term:
-/// - return an empty collection if the term is to be excluded from analysis;
-/// - return multiple terms if the term is split; and/or
-/// - return modified term(s), such as applying a stemmer algorithm.
-
-/// `SentenceSplitter` - A splitter function that returns a list of sentences from [source].
-///
-/// In English, the [source] text is split at sentence endings marks such as
-/// periods, question marks and exclamation marks. If the [source] contains
-/// paragraph ending marks it should also be split at these.
-///
-/// The sentence splitter should avoid splitting after abbreviations,
-/// which may end with period marks.
-
-/// An alias for String, when used in the context of a field or meta data field
-/// in the `corpus`. Represents the name of the field/zone.
-
-/// An alias for String, used in the context of a word, term or phrase present
-/// in a text source, document or query.
-
-/// Phrase: An alias for `List<String>` when used in the context of the terms of a
-/// phrase split to an ordered list of terms.
-
-/// k-Gram: An alias for String, used in the context of a sequence of k characters
-/// from a term.
-///
-/// A k-gram can start with "$", dentoting the start of the term, and end with
-/// "$", denoting the end of the term.
-///
-/// The 3-grams for "castle" are { $ca, cas, ast, stl, tle, le$ }
-
-/// `PhraseExtractor` - a splitter function that returns an ordered collection of keyword phrases
-/// from text.
-///
-/// The text is split at punctuation, line endings and stop-words, resulting
-/// in an ordered collection of term sequences of varying length.
-
 /// Alias for `Map<String, Set<String>>`.
 ///
 /// A hashmap of k-Gram to Set<term>, where the value is the set of unique
@@ -107,10 +59,7 @@ typedef PhraseSplitter = Future<List<String>> Function(String source,
 ///
 /// Returns a List<[Token]>.
 typedef Tokenizer = Future<List<Token>> Function(String source,
-    {NGramRange? nGramRange,
-    bool preserveCase,
-    String? zone,
-    TokenFilter? tokenFilter});
+    {NGramRange? nGramRange, String? zone, TokenFilter? tokenFilter});
 
 /// Type definition of a function that returns a collection of [Token] from
 /// the [zones] in a JSON [document].
@@ -129,7 +78,6 @@ typedef Tokenizer = Future<List<Token>> Function(String source,
 typedef JsonTokenizer = Future<List<Token>> Function(
     Map<String, dynamic> document,
     {NGramRange? nGramRange,
-    bool preserveCase,
     Iterable<String>? zones,
     TokenFilter? tokenFilter});